requirements.txt
torch
torchvision
scikit-image
numpy
opencv-python
matplotlib
scikit-learn
shap
pillow